Implement q_remove_head function
Using `list_first_entry`, retrieve the address of the first node. Then, use `list_del_init` to remove the node and make its head and tail point to itself.
Finally, copy the value of the removed node to `sp`, and when copying the string to `sp`, allocate space for the null terminator.