Implement q_delete_dup
Use `list_for_each_entry_safe` to traverse every node and obtain the current and next addresses of `element_t`.
Check if the next node of the current node points to `head`.
Use `strcmp` to compare the values of the current node and the next node.
If they are the same, remove and free the memory space of the duplicate node using `list_del_init` and `q_release_element`.
Declare a variable `flag` to ensure that if the strings are the same, another duplicate node will be removed in the next iteration using the same method.