Implement q_swap function
Declare a pointer `cur` pointing to `head`, and then declare two pointers pointing to the first and second nodes respectively.
In each iteration, swap the pointers pairwise, and before the end of the loop, move both pointers two nodes back until one of them points to `head`.
It's important to pay attention to the order of updating the linked pointers.
Incorrect update logic may lead to a distorted linked structure, such as processing nodes redundantly or forming a cyclic linked list.