Implement q_reverseK and q_sort function
In q_reverseK:
Reverse each group of k nodes one by one and iterate through each group.

Use `q_size` to record the overall length of the queue to ensure that the last remaining nodes, which are fewer than k, are not reversed,
and use indirect pointer `reverse_node` to record the address of `cur->next`, while `tmp` is used to record the head of each group.

Apply the method of `q_reverse` to reverse the current group.
At this point, `cur` will point to the last node of the group, and assign `cur->next` to `cur`,
so `cur` will point to the first node of the next group.

Therefore, the next iteration will start from the node pointed to by `*reverse_node` and reverse each node to the front of the group,
achieving the reversal.

In q_sort:
In the `trace-15-perf.cmd` file, it mentions that the test for sorting requires a time complexity of O(nlogn). Using sorting algorithms with a time complexity of O(n^2), such as bubble sort, would cause the test to fail. Therefore, merge sort was chosen to meet this expectation.

Using three functions to illustrate:

In `q_sort`:
Before executing `merge_sort`, the `head` must be disconnected to make it a singly linked list to avoid infinite loops.
After `merge_sort`, an additional iteration through the sorted linked list is required. Since the list is now a singly linked list after sorting, the `prev` pointers need to be updated. Additionally, the tail and head of the queue must be reconnected to restore the circular doubly-linked list structure to maintain the `list_head` type.

In `merge_sort`:
If the linked list is empty or contains only one node, it returns directly.
The middle node is found using the fast and slow pointer method, then `merge_list` is called to merge the linked lists starting from the first node and the middle node into one sorted linked list.
Setting `slow->next = NULL;` is necessary to point the end of the linked list starting from the first node to NULL.

In `merge_list`:
Using `list_entry` allows obtaining a pointer to the `element_t` structure in the `struct list_head` structure pointed to by `left`. Then, the values of `element_t` are compared, and the one with the smaller value is placed in the new queue. The `ptr` is then moved to the next position. After the iteration is completed, the remaining nodes are added to the queue.
Using the technique of indirect pointers avoids creating additional node memory during the queue merging process.